+++ src/components/table-ajax/table-ajax.js
@@ -92,7 +92,25 @@ class TableAjax extends Table {
      * @property totalRecords
      * @type {String}
      */
-    totalRecords: '0'
+    totalRecords: '0',
+
+    /**
+     * Sorting
+     * either 'asc' or 'desc' order
+     *
+     * @property sortOrder
+     * @type {String}
+     */
+    sortOrder: this.props.sortOrder || '',
+    /**
+     * Sorting
+     * column name to sort
+     *
+     * @property sortedColumn
+     * @type {String}
+     */
+    sortedColumn: this.props.sortedColumn || ''
+
   };
 
   /**
@@ -118,6 +136,33 @@ class TableAjax extends Table {
     this.resizeTable();
   }
 
+  static childContextTypes = {
+    /**
+     * Defines a context object for child components of the table component.
+     * https://facebook.github.io/react/docs/context.html
+     *
+     * @property table
+     * @type {Object}
+     */
+    onSort: React.PropTypes.func,
+    sortedColumn: React.PropTypes.string,
+    sortOrder: React.PropTypes.string
+  }
+
+  /**
+   * Returns table object to child components.
+   *
+   * @method getChildContext
+   * @return {void}
+   */
+  getChildContext = () => {
+    return {
+      onSort: this.onSort,
+      sortedColumn: this.sortedColumn,
+      sortOrder: this.sortOrder
+    };
+  }
+
   /**
    * Get pageSize for table
    * @override
@@ -127,6 +172,26 @@ class TableAjax extends Table {
    */
   get pageSize() {
     return this.state.pageSize;
+  }
+
+  /**
+   * Returns the currently sorted column.
+   *
+   * @method sortedColumn
+   * @return {String}
+   */
+  get sortedColumn() {
+    return this.state.sortedColumn;
+  }
+
+  /**
+   * Returns the current sort order.
+   *
+   * @method sortOrder
+   * @return {String}
+   */
+  get sortOrder() {
+    return this.state.sortOrder;
   }
 
   /**
@@ -143,21 +208,20 @@ class TableAjax extends Table {
 
     this.setState({
       currentPage: options.currentPage,
-      pageSize: options.pageSize
+      pageSize: options.pageSize,
+      sortOrder: options.sortOrder,
+      sortedColumn: options.sortedColumn
     });
 
     this.stopTimeout();
     this.timeout = setTimeout(() => {
       Request
         .get(this.props.path)
+        .set('Accept', 'application/json')
         .query(this.queryParams(element, options))
         .end((err, response) => {
-          if (!err) {
-            let data = response.body.data[0];
-            this.props.onChange(data);
-            this.setState({totalRecords: String(data.records)});
-            if (resetHeight) { this.resetTableHeight(); }
-          }
+          this.handleResponse(err, response);
+          if (resetHeight) { this.resetTableHeight(); }
         });
     }, timeout);
   }
@@ -175,6 +239,21 @@ class TableAjax extends Table {
   }
 
   /**
+   * Handles what happens with response.
+   *
+   * @method handlerResponse
+   * @param {Object} err
+   * @param {Object} response
+   */
+  handleResponse = (err, response) => {
+    if (!err) {
+      let data = response.body;
+      this.props.onChange(data);
+      this.setState({ totalRecords: String(data.records) });
+    }
+  }
+
+  /**
    * Formatted params for server request
    *
    * @method queryParams
@@ -186,6 +265,8 @@ class TableAjax extends Table {
     let query = options.filter || {};
     query.page = options.currentPage;
     query.rows = options.pageSize;
+    if (options.sortOrder) { query.sord = options.sortOrder; }
+    if (options.sortedColumn) { query.sidx = options.sortedColumn; }
     return serialize(query);
   }
 
@@ -200,7 +281,9 @@ class TableAjax extends Table {
     return {
       currentPage: this.state.currentPage,
       filter: props.filter ? props.filter.toJS() : {},
-      pageSize: this.state.pageSize
+      pageSize: this.state.pageSize,
+      sortedColumn: this.state.sortedColumn,
+      sortOrder: this.state.sortOrder
     };
   }