+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## Bug Fixes
 
 * TableCell and TableHeader can receive additional props.
+* Inputs no longer render a label if the input has no name or label props.
 
 # 0.6.0